don't reinstall deps

+++ share/scripts/install-cmake.sh
@@ -8,14 +8,18 @@ INSTALL_PREFIX=${HOME}/local/$(uname)
 mkdir -p ${TMP_DIR}
 mkdir -p ${INSTALL_PREFIX}
 
-if ! [ -f "${TMP_DIR}/`basename $URL`" ]; then
-    wget ${URL} -O ${TMP_DIR}/`basename $URL`
+if ! [ -f "${INSTALL_PREFIX}/bin/cmake" ]; then
+    if ! [ -f "${TMP_DIR}/`basename $URL`" ]; then
+        wget ${URL} -O ${TMP_DIR}/`basename $URL`
+    fi
+
+    chmod +x ${TMP_DIR}/`basename $URL`
+
+    cd ${TMP_DIR}
+
+    ${TMP_DIR}/`basename $URL` --prefix=${INSTALL_PREFIX} --exclude-subdir || exit $?
+
 fi
 
-chmod +x ${TMP_DIR}/`basename $URL`
-
-cd ${TMP_DIR}
-
-${TMP_DIR}/`basename $URL` --prefix=${INSTALL_PREFIX} --exclude-subdir || exit $?
-
 exit $?
+
